move content component to its own file
